rename ButtonAction interface to indicate specifying generic event type
